Create cstr-reactor-system.jl
Simulation of a cstr reactor through a big system of non-linear equations using the newton-raphson method. The problem was extract from the book: A Step by Step Approach to the Modeling of Chemical Engineering Processes: Using Excel for Simulation. Due to the highly non-linear behavior of the system, numerical solution using Excel was unfeasible and a Julia programming language solution was implemented.